Manually Bootstrapping AngularJS
(this allows calls to native org.apache.cordova.plugin.console to be
made anywhere in the code, along with calls to all other Cordova plugins)

We are now bootstrapping AngularJS only after the 'deviceready' event
fires, which keeps our app from attempting to call native device
functions before the device is ready to receive those calls.

The reason this change was necessary is because Randy Johnson (from CIMA)
was having intermittent success getting console.log() messages to appear
in the XCode console window.  Specifically, he was attempting to log
console messages from within the factories.js source, but that piece of
code (and probably other pieces too) was running before the device was
ready, so console.log() was unavailable.

Side Effects:  There are no known negative consequences of this change.
However, the benefits of this change include the ability for developers
to now make calls to all native device (Cordova) APIs [not just to
console.log()] from anywhere in the
code.